services/liquidation-bot/requirements.txt
web3==6.15.0
python-dotenv==1.0.1
eth-abi==5.0.0
eth-account==0.11.0
requests==2.31.0
schedule==1.2.1
